webapi/tct-webmessaging-w3c-tests/webmessaging-py/w3c/postMessage_arrays.htm

<!DOCTYPE html>
<html>
<head>
<title> postMessage with arrays </title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/server.js?pipe=sub"></script>
</head>
<body>
<div id=log></div>

<script>
    var DATA = [1,2,3,4,5,6,7,8];
    var TYPE = "object";
    var TARGET;

    function PostMessageTest()
    {
        TARGET = document.querySelector("iframe");
        TARGET.contentWindow.postMessage(DATA, "*");
    }
</script>

<div style="display:none">
    <iframe width="70%" onload="PostMessageTest()" src=""></iframe>
</div>

<script>
    setup({timeout:500});

    path = "http://" + __SERVER__NAME + ":" + __CORS__PORT + "/tests/webmessaging/w3c/support/ChildWindowPostMessage.htm";
    TARGET.src = path;

    var description = "Test Description: Messages can be structured objects, e.g., arrays.";

    var t = async_test(description);

    window.onmessage = t.step_func(function(e)
    {
        assert_equals(typeof(e.data), TYPE);
        assert_array_equals(e.data, DATA);
        t.done();
    });
</script>
</body>
</html>
